+++ src/main/java/com/study/sns/domain/user/model/constant/UserStatus.java
@@ -1,4 +1,4 @@
-package com.study.sns.model.constant;
+package com.study.sns.domain.user.model.constant;
 
 public enum UserStatus {
     ACTIVE,
